Text Detection and Extraction/requirements.txt
flask
psycopg2-binary
flask-socketio
python-dotenv
gunicorn